HTCONDOR-1003 Create a global cm that only knows about schedds
