+++ src/condor_negotiator.V6/matchmaker.cpp
@@ -3024,6 +3024,14 @@ obtainAdsFromCollector (
 				ad->AssignExpr(ATTR_SLOT_WEIGHT, slotWeightStr);
 			}
 
+			if (param_boolean("NEGOTIATOR_SCHEDDS_ARE_SUBMITTERS", false)) {
+				std::string scheddName;
+				if (!ad->LookupString(ATTR_SCHEDD_NAME, scheddName)) {
+					dprintf(D_ALWAYS, "WARNING: slot ad %s is missing a schedd name, skipping\n",remoteHost);
+					continue;
+				}
+				ad->Assign(ATTR_REMOTE_USER, scheddName);
+			}
 			OptimizeMachineAdForMatchmaking( ad );
 
 			startdAds.Insert(ad);
@@ -3040,6 +3048,16 @@ obtainAdsFromCollector (
 			if ( !IsValidSubmitterName( subname.c_str() ) ) {
 				dprintf( D_ALWAYS, "WARNING: ignoring submitter ad with invalid name: %s\n", subname.c_str() );
 				continue;
+			}
+
+			if (param_boolean("NEGOTIATOR_SCHEDDS_ARE_SUBMITTERS", false)) {
+				std::string scheddName;
+				if (!ad->LookupString(ATTR_SCHEDD_NAME, scheddName)) {
+					dprintf(D_ALWAYS, "WARNING: Submitter Ad %s is missing a schedd name, skipping\n",subname.c_str());
+					continue;
+				}
+				ad->Assign(ATTR_NAME, scheddName);
+				subname = scheddName;
 			}
 
 			if (!TransformSubmitterAd(*ad)) {
@@ -3061,7 +3079,7 @@ obtainAdsFromCollector (
                 continue;
             }
 
-			submitterNames.insert(std::string(subname.c_str()));
+			submitterNames.insert(subname);
 
     		ad->Assign(ATTR_TOTAL_TIME_IN_CYCLE, 0);
 
@@ -3716,6 +3734,11 @@ Matchmaker::startNegotiateProtocol(const std::string &submitter, const ClassAd &
 		negotiate_ad.InsertAttr(ATTR_AUTO_CLUSTER_ATTRS, job_attr_references ? job_attr_references : "");
 		// Tell the schedd a submitter tag value (used for flocking levels)
 		negotiate_ad.InsertAttr(ATTR_SUBMITTER_TAG, submitter_tag.c_str());
+
+		if (param_boolean("NEGOTIATOR_SCHEDDS_ARE_SUBMITTERS", false)) {
+			negotiate_ad.InsertAttr(ATTR_NEGOTIATOR_SCHEDDS_ARE_SUBMITTERS, true);
+		}
+
 		if (!putClassAd(sock, negotiate_ad))
 		{
 			dprintf(D_ALWAYS, "    Failed to send negotiation header to %s\n",
@@ -4793,106 +4816,6 @@ matchmakingAlgorithm(const char *submitterName, const char *scheddAddr, ClassAd 
 	return bestSoFar;
 }
 
-class NotifyStartdOfMatchHandler {
-public:
-	std::string m_startdName;
-	std::string  m_startdAddr;
-	std::string  m_claim_id;
-	int m_timeout;
-	DCStartd m_startd;
-	bool m_nonblocking;
-
-	NotifyStartdOfMatchHandler(char const *startdName,char const *startdAddr,int timeout,char const *claim_id,bool nonblocking):
-		
-		m_startdName(startdName),
-		m_startdAddr(startdAddr),
-		m_claim_id(claim_id),
-		m_timeout(timeout),
-		m_startd(startdAddr),
-		m_nonblocking(nonblocking) {}
-
-	static void startCommandCallback(bool success,Sock *sock,CondorError * /*errstack*/,
-		const std::string & /*trust_domain*/, bool /*should_try_token_request*/, void *misc_data)
-	{
-		NotifyStartdOfMatchHandler *self = (NotifyStartdOfMatchHandler *)misc_data;
-		ASSERT(misc_data);
-
-		if(!success) {
-			dprintf (D_ALWAYS,"      Failed to initiate socket to send MATCH_INFO to %s\n",
-					 self->m_startdName.c_str());
-		}
-		else {
-			self->WriteMatchInfo(sock);
-		}
-		if(sock) {
-			delete sock;
-		}
-		delete self;
-	}
-
-	bool WriteMatchInfo(Sock *sock) const
-	{
-		ClaimIdParser idp( m_claim_id.c_str() );
-		ASSERT(sock);
-
-		// pass the startd MATCH_INFO and claim id string
-		dprintf (D_FULLDEBUG, "      Sending MATCH_INFO/claim id to %s\n",
-		         m_startdName.c_str());
-		dprintf (D_FULLDEBUG, "      (Claim ID is \"%s\" )\n",
-		         idp.publicClaimId() );
-
-		if ( !sock->put_secret (m_claim_id.c_str()) ||
-			 !sock->end_of_message())
-		{
-			dprintf (D_ALWAYS,
-			        "      Could not send MATCH_INFO/claim id to %s\n",
-			        m_startdName.c_str() );
-			dprintf (D_FULLDEBUG,
-			        "      (Claim ID is \"%s\")\n",
-			        idp.publicClaimId() );
-			return false;
-		}
-		return true;
-	}
-
-	bool startCommand()
-	{
-		dprintf (D_FULLDEBUG, "      Connecting to startd %s at %s\n",
-					m_startdName.c_str(), m_startdAddr.c_str());
-
-		if(!m_nonblocking) {
-			Stream::stream_type st = m_startd.hasUDPCommandPort() ? Stream::safe_sock : Stream::reli_sock;
-			Sock *sock =  m_startd.startCommand(MATCH_INFO,st,m_timeout);
-			bool result = false;
-			if(!sock) {
-				dprintf (D_ALWAYS,"      Failed to initiate socket (blocking mode) to send MATCH_INFO to %s\n",
-						 m_startdName.c_str());
-			}
-			else {
-				result = WriteMatchInfo(sock);
-			}
-			if(sock) {
-				delete sock;
-			}
-			delete this;
-			return result;
-		}
-
-		Stream::stream_type st = m_startd.hasUDPCommandPort() ? Stream::safe_sock : Stream::reli_sock;
-		m_startd.startCommand_nonblocking (
-			MATCH_INFO,
-			st,
-			m_timeout,
-			NULL,
-			NotifyStartdOfMatchHandler::startCommandCallback,
-			this);
-
-			// Since this is nonblocking, we cannot give any immediate
-			// feedback on whether the message to the startd succeeds.
-		return true;
-	}
-};
-
 void Matchmaker::
 insertNegotiatorMatchExprs( ClassAdListDoesNotDeleteAds &cal )
 {
@@ -5133,6 +5056,10 @@ matchmakingProtocol (ClassAd &request, ClassAd *offer,
 	// used by schedd_negotiate.cpp when resource request lists are being used
 	offer->Assign(ATTR_RESOURCE_REQUEST_CLUSTER,cluster);
 	offer->Assign(ATTR_RESOURCE_REQUEST_PROC,proc);
+
+	if (param_boolean("NEGOTIATOR_SCHEDDS_ARE_SUBMITTERS", false)) {
+		offer->InsertAttr(ATTR_NEGOTIATOR_SCHEDDS_ARE_SUBMITTERS, true);
+	}
 
 	// ---- real matchmaking protocol begins ----
 	// 1.  contact the startd
